% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks-advanced.R
\name{add_callback}
\alias{add_callback}
\title{Add a callback to a Dash app}
\usage{
add_callback(app, outputs, params, callback)
}
\arguments{
\item{app}{A dash application created with \code{\link[=dash_app]{dash_app()}}.}
}
\description{
Add a callback to a Dash app
}
