% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{checklist}
\alias{checklist}
\title{checklist component 
Checklist is a component that encapsulates several checkboxes. The values and labels of the checklist is specified in the `options` property and the checked items are specified with the `values` property. Each checkbox is rendered as an input with a surrounding label.}
\usage{
checklist(id = NULL, inputStyle = NULL, inputClassName = NULL,
  labelStyle = NULL, labelClassName = NULL, options = NULL,
  values = NULL, fireEvent = NULL, setProps = NULL, dashEvents = NULL)
}
\arguments{
\item{id}{}

\item{inputStyle}{The style of the <input> checkbox element}

\item{inputClassName}{The class of the <input> checkbox element}

\item{labelStyle}{The style of the <label> that wraps the checkbox input  and the option's label}

\item{labelClassName}{The class of the <label> that wraps the checkbox input  and the option's label}

\item{options}{An array of options}

\item{values}{The currently selected value}

\item{fireEvent}{Dash-assigned callback that gets fired when the checkbox item gets selected.}

\item{setProps}{Dash-assigned callback that gets fired when the value changes.}

\item{dashEvents}{}
}
\description{
checklist component 
Checklist is a component that encapsulates several checkboxes. The values and labels of the checklist is specified in the `options` property and the checked items are specified with the `values` property. Each checkbox is rendered as an input with a surrounding label.
}
