% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-core.R
\name{core_checklist}
\alias{core_checklist}
\title{Checklist core component}
\usage{
core_checklist(..., id = NULL, options = NULL, values = NULL,
  className = NULL, style = NULL, inputStyle = NULL,
  inputClassName = NULL, labelStyle = NULL, labelClassName = NULL,
  fireEvent = NULL, setProps = NULL, dashEvents = NULL)
}
\arguments{
\item{...}{children of the component. Either other components or \link[htmltools:tags]{htmltools::tags} are allowed.}

\item{id}{}

\item{options}{An array of options}

\item{values}{The currently selected value}

\item{className}{The class of the container (div)}

\item{style}{The style of the container (div)}

\item{inputStyle}{The style of the <input> checkbox element}

\item{inputClassName}{The class of the <input> checkbox element}

\item{labelStyle}{The style of the <label> that wraps the checkbox input  and the option's label}

\item{labelClassName}{The class of the <label> that wraps the checkbox input  and the option's label}

\item{fireEvent}{Dash-assigned callback that gets fired when the checkbox item gets selected.}

\item{setProps}{Dash-assigned callback that gets fired when the value changes.}

\item{dashEvents}{}
}
\description{
Checklist is a component that encapsulates several checkboxes. The values and labels of the checklist is specified in the \code{options} property and the checked items are specified with the \code{values} property. Each checkbox is rendered as an input with a surrounding label.
}
