% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-core.R
\name{core_dropdown}
\alias{core_dropdown}
\title{Dropdown core component}
\usage{
core_dropdown(..., id = NULL, options = NULL, value = NULL,
  className = NULL, clearable = NULL, disabled = NULL, multi = NULL,
  placeholder = NULL, searchable = NULL, setProps = NULL,
  dashEvents = NULL)
}
\arguments{
\item{...}{children of the component. Either other components or \link[htmltools:tags]{htmltools::tags} are allowed.}

\item{id}{}

\item{options}{An array of options}

\item{value}{The value of the input. If \code{multi} is false (the default) then value is just a string that corresponds to the values provided in the \code{options} property. If \code{multi} is true, then multiple values can be selected at once, and \code{value} is an array of items with values corresponding to those in the \code{options} prop.}

\item{className}{className of the dropdown element}

\item{clearable}{Whether or not the dropdown is "clearable", that is, whether or not a small "x" appears on the right of the dropdown that removes the selected value.}

\item{disabled}{If true, the option is disabled}

\item{multi}{If true, the user can select multiple values}

\item{placeholder}{The grey, default text shown when no option is selected}

\item{searchable}{Whether to enable the searching feature or not}

\item{setProps}{Dash-assigned callback that gets fired when the input changes}

\item{dashEvents}{}
}
\description{
Dropdown is an interactive dropdown element for selecting one or more items. The values and labels of the dropdown items are specified in the \code{options} property and the selected item(s) are specified with the \code{value} property.  Use a dropdown when you have many options (more than 5) or when you are constrained for space. Otherwise, you can use RadioItems or a Checklist, which have the benefit of showing the users all of the items at once.
}
