% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-core.R
\name{core_interval}
\alias{core_interval}
\title{Interval core component}
\usage{
core_interval(..., id = NULL, interval = 1000, fireEvent = NULL,
  setProps = NULL, dashEvents = NULL)
}
\arguments{
\item{...}{children of the component. Either other components or \link[htmltools:tags]{htmltools::tags} are allowed.}

\item{id}{}

\item{interval}{This component will fire an event every \code{interval} milliseconds with the event name \code{setInterval}}

\item{fireEvent}{Dash assigned callback}

\item{setProps}{Dash assigned callback}

\item{dashEvents}{}
}
\description{
A component that repeatedly fires an event ("interval") with a fixed time delay between each event. Interval is good for triggering a component on a recurring basis. The time delay is set with the property "interval" in milliseconds.
}
