% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-core.R
\name{core_location}
\alias{core_location}
\title{Location core component}
\usage{
core_location(..., id = NULL, pathname = NULL, refresh = TRUE)
}
\arguments{
\item{...}{children of the component. Either other components or \link[htmltools:tags]{htmltools::tags} are allowed.}

\item{id}{(required)}

\item{pathname}{}

\item{refresh}{}
}
