% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-core.R
\name{core_markdown}
\alias{core_markdown}
\title{Markdown core component}
\usage{
core_markdown(..., id = NULL, className = NULL, containerProps = NULL)
}
\arguments{
\item{...}{children of the component. Either other components or \link[htmltools:tags]{htmltools::tags} are allowed.}

\item{id}{}

\item{className}{Class name of the container element}

\item{containerProps}{An object containing custom element props to put on the container element such as id or style}
}
\description{
A component that renders Markdown text as specified by the CommonMark spec.
}
