% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dash.R
\docType{class}
\name{Dash}
\alias{Dash}
\title{Generate a Dash server}
\description{
The Dash generator inherits from the \link[fiery:Fire]{fiery::Fire} class.
}
\section{Initialization}{

A new 'Dash'-object is initialized using the \code{new()} method on the generator:

\strong{Usage}
\tabular{l}{
\code{app <- Dash$new(name = "dash", url_base_pathname = "/", csrf_protect = TRUE)}
}

\strong{Arguments}
\tabular{lll}{
\code{host} \tab  \tab A string overriding the default host (see the \emph{Fields} section below)\cr
\code{port} \tab  \tab An integer overriding the default port (see the \emph{Fields} section below)
}
}

\section{Fields}{

\describe{
\item{\code{name}}{A name for the dash server.}
\item{\code{url_base_pathname}}{a path for prepending to dash endpoints.}
\item{`csrf_protect``}{prevent cross-site request forgery (CSRF)}
}
}

\section{Methods}{

\describe{
\item{}
}
}

\examples{


# hello dash!
app <- dasher::Dash$new()
\dontrun{
app$run_server()
}


# use the layout_set method to define your UI
app$layout_set(
  html_h2("Hello Dash"),
  "Here is a slider",
  core_slider()
)
# retrieve the layout with layout_get()
app$layout_get()


}
\keyword{datasets}
