% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dash_filter_null}
\alias{dash_filter_null}
\title{Filter NULL values from component properties}
\usage{
dash_filter_null(x)
}
\arguments{
\item{x}{A list of component properties.}
}
\description{
This helper function is used by Dash to filter NULL values before
serializing component properties and values into JSON.
}
