% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dash-package.R
\docType{package}
\name{dasher-package}
\alias{dasher}
\alias{dasher-package}
\title{An implementation of the dash protocol for authoring reactive web applications}
\description{
Coming soon
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/plotly/dasher}
  \item Report bugs at \url{https://github.com/plotly/dasher/issues}
}

}
\author{
\strong{Maintainer}: Carson Sievert \email{cpsievert1@gmail.com}

}
