% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashBootstrapComponents.R
\docType{data}
\name{dbcIcons}
\alias{dbcIcons}
\title{dbcIcons}
\format{
An object of class \code{list} of length 2.
}
\usage{
dbcIcons
}
\description{
A list of contextually colored icon styles that can be added to Dash Bootstrap Components.
}
\keyword{datasets}
