% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashBootstrapComponents.R
\docType{data}
\name{dbcThemes}
\alias{dbcThemes}
\title{dbcThemes}
\format{
An object of class \code{list} of length 27.
}
\usage{
dbcThemes
}
\description{
Externally hosted themes that can be passed to the Dash app with \code{app \%>\% add_stylesheet()}.
}
\keyword{datasets}
