% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{dependencies}
\alias{dependencies}
\alias{output}
\alias{input}
\alias{state}
\title{Input/Output/State definitions}
\usage{
output(component_id = NULL, component_property = "children")

input(component_id = NULL, component_property = "value")

state(component_id = NULL, component_property = "value")
}
\description{
Use in conjunction with the \code{callback()} method from the \link[dasher:Dash]{dasher::Dash} class.
}
\author{
Carson Sievert
}
