% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_list.R
\name{df_to_list}
\alias{df_to_list}
\title{df_to_list}
\usage{
df_to_list(df)
}
\arguments{
\item{df}{A data frame object to be coerced into a list of lists for DataTable.}
}
\description{
Helper function to convert a dataframe into the nested list format required for input into Dash DataTable.
}
