% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{dropdown}
\alias{dropdown}
\title{dropdown component 
Dropdown is an interactive dropdown element for selecting one or more items. The values and labels of the dropdown items are specified in the `options` property and the selected item(s) are specified with the `value` property.  Use a dropdown when you have many options (more than 5) or when you are constrained for space. Otherwise, you can use RadioItems or a Checklist, which have the benefit of showing the users all of the items at once.}
\usage{
dropdown(id = NULL, className = NULL, disabled = NULL, multi = NULL,
  options = NULL, placeholder = NULL, value = NULL, setProps = NULL,
  dashEvents = NULL)
}
\arguments{
\item{id}{}

\item{className}{}

\item{disabled}{If true, the option is disabled}

\item{multi}{If true, the user can select multiple values}

\item{options}{}

\item{placeholder}{The grey, default text shown when no option is selected}

\item{value}{The value of the input. If `multi` is false (the default) then value is just a string that corresponds to the values provided in the `options` property. If `multi` is true, then multiple values can be selected at once, and `value` is an array of items with values corresponding to those in the `options` prop.}

\item{setProps}{Dash-assigned callback that gets fired when the input changes}

\item{dashEvents}{}
}
\description{
dropdown component 
Dropdown is an interactive dropdown element for selecting one or more items. The values and labels of the dropdown items are specified in the `options` property and the selected item(s) are specified with the `value` property.  Use a dropdown when you have many options (more than 5) or when you are constrained for space. Otherwise, you can use RadioItems or a Checklist, which have the benefit of showing the users all of the items at once.
}
