% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-output-value.R
\name{format_output_value}
\alias{format_output_value}
\title{Format callback output}
\usage{
format_output_value(x, user_deps, ...)
}
\arguments{
\item{x}{the output value from a callback function}

\item{user_deps}{user specified HTML dependencies}

\item{...}{additional arugments. Currently unused.}
}
\description{
This function is not intended for direct use by users, but has been exported
for development purposes.
}
\details{
This (generic) function is called on the output value of callback
function(s) -- effectively providing a way for developers to provide
a bridge between an R object and (an existing) suitable dash component.
}
