% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{graph}
\alias{graph}
\title{graph component}
\usage{
graph(id = NULL, clickData = NULL, hoverData = NULL,
  clear_on_unhover = NULL, selectedData = NULL, relayoutData = NULL,
  figure = NULL, style = NULL, animate = NULL, animation_options = NULL,
  config = NULL, dashEvents = NULL, setProps = NULL,
  dashFireEvent = NULL)
}
\arguments{
\item{id}{(required)}

\item{clickData}{Data from latest click event}

\item{hoverData}{Data from latest hover event}

\item{clear_on_unhover}{If True, `clear_on_unhover` will clear the `hoverData` property when the user "unhovers" from a point. If False, then the `hoverData` property will be equal to the data from the last point that was hovered over.}

\item{selectedData}{Data from latest select event}

\item{relayoutData}{Data from latest relayout event which occurs when the user zooms or pans on the plot}

\item{figure}{Plotly `figure` object. See schema: https://plot.ly/javascript/reference}

\item{style}{Generic style overrides on the plot div}

\item{animate}{Beta: If true, animate between updates using plotly.js's `animate` function}

\item{animation_options}{Beta: Object containing animation settings. Only applies if `animate` is `true`}

\item{config}{Plotly.js config options. See https://plot.ly/javascript/configuration-options/ for more info.}

\item{dashEvents}{}

\item{setProps}{Function that updates the state tree.}

\item{dashFireEvent}{Function that fires events}
}
\description{

}
