% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heroku.R
\name{heroku_app_deploy}
\alias{heroku_app_deploy}
\title{Create (and deploy) dashR app to Heroku}
\usage{
heroku_app_deploy(name = NULL)
}
\arguments{
\item{name}{a name for (a new) Heroku app.}
}
\description{
This function tries to deploy a dashR app in the current directory to Heroku.
}
\details{
dashR app to Heroku. It can also deploy the template app
if your system is configured properly (see details).

To deploy, you need the \href{https://devcenter.heroku.com/articles/heroku-cli}{Heroku CLI}
and successfully run \code{heroku login} ahead of time. You'll also need access to the \code{git}
system command.
}
\references{
\url{https://github.com/plotly/dashR/tree/master/inst/heroku}
}
\author{
Carson Sievert
}
