% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heroku.R
\name{heroku_app_template}
\alias{heroku_app_template}
\title{Summon a template for deploying dashR to heroku}
\usage{
heroku_app_template(overwrite = FALSE)
}
\description{
This function copies files for Heroku deployment of a dashR app to the
current directory.
dashR app to Heroku via \href{https://github.com/virtualstaticvoid/heroku-buildpack-r/tree/heroku-16}{the R buildpack}.
}
\author{
Carson Sievert
}
