% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-html.R
\name{html_audio}
\alias{html_audio}
\title{audio html component}
\usage{
html_audio(..., id = NULL, n_clicks = NULL, autoPlay = NULL,
  controls = NULL, crossOrigin = NULL, loop = NULL, preload = NULL,
  src = NULL, accessKey = NULL, className = NULL,
  contentEditable = NULL, contextMenu = NULL, dir = NULL,
  draggable = NULL, hidden = NULL, lang = NULL, spellCheck = NULL,
  style = NULL, tabIndex = NULL, title = NULL, fireEvent = NULL,
  dashEvents = NULL)
}
\arguments{
\item{...}{children of the component. Either other components or \code{\link[htmltools:tags]{htmltools::tags()}} are allowed.}

\item{id}{The ID of this component, used to identify dash components in callbacks. The ID needs to be unique across all of the components in an app.}

\item{n_clicks}{An integer that represents the number of times that this element has been clicked on.}

\item{autoPlay}{The audio or video should play as soon as possible.}

\item{controls}{Indicates whether the browser should show playback controls to the user.}

\item{crossOrigin}{How the element handles cross-origin requests}

\item{loop}{Indicates whether the media should start playing from the start when it's finished.}

\item{preload}{Indicates whether the whole resource, parts of it or nothing should be preloaded.}

\item{src}{The URL of the embeddable content.}

\item{accessKey}{Defines a keyboard shortcut to activate or add focus to the element.}

\item{className}{Often used with CSS to style elements with common properties.}

\item{contentEditable}{Indicates whether the element's content is editable.}

\item{contextMenu}{Defines the ID of a <menu> element which will serve as the element's context menu.}

\item{dir}{Defines the text direction. Allowed values are ltr (Left-To-Right) or rtl (Right-To-Left)}

\item{draggable}{Defines whether the element can be dragged.}

\item{hidden}{Prevents rendering of given element, while keeping child elements, e.g. script elements, active.}

\item{lang}{Defines the language used in the element.}

\item{spellCheck}{Indicates whether spell checking is allowed for the element.}

\item{style}{Defines CSS styles which will override styles previously set.}

\item{tabIndex}{Overrides the browser's default tab order and follows the one specified instead.}

\item{title}{Text to be displayed in a tooltip when hovering over the element.}

\item{fireEvent}{A callback for firing events to dash.}

\item{dashEvents}{}
}
\description{
See \url{https://developer.mozilla.org/en-US/docs/Web/HTML/Element/audio}
}
