% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-html.R
\name{html_select}
\alias{html_select}
\title{select html component}
\usage{
html_select(..., id = NULL, n_clicks = NULL, autoFocus = NULL,
  disabled = NULL, form = NULL, multiple = NULL, name = NULL,
  required = NULL, size = NULL, accessKey = NULL, className = NULL,
  contentEditable = NULL, contextMenu = NULL, dir = NULL,
  draggable = NULL, hidden = NULL, lang = NULL, spellCheck = NULL,
  style = NULL, tabIndex = NULL, title = NULL, fireEvent = NULL,
  dashEvents = NULL)
}
\arguments{
\item{...}{children of the component. Either other components or \code{\link[htmltools:tags]{htmltools::tags()}} are allowed.}

\item{id}{The ID of this component, used to identify dash components in callbacks. The ID needs to be unique across all of the components in an app.}

\item{n_clicks}{An integer that represents the number of times that this element has been clicked on.}

\item{autoFocus}{The element should be automatically focused after the page loaded.}

\item{disabled}{Indicates whether the user can interact with the element.}

\item{form}{Indicates the form that is the owner of the element.}

\item{multiple}{Indicates whether multiple values can be entered in an input of the type email or file.}

\item{name}{Name of the element. For example used by the server to identify the fields in form submits.}

\item{required}{Indicates whether this element is required to fill out or not.}

\item{size}{Defines the width of the element (in pixels). If the element's type attribute is text or password then it's the number of characters.}

\item{accessKey}{Defines a keyboard shortcut to activate or add focus to the element.}

\item{className}{Often used with CSS to style elements with common properties.}

\item{contentEditable}{Indicates whether the element's content is editable.}

\item{contextMenu}{Defines the ID of a <menu> element which will serve as the element's context menu.}

\item{dir}{Defines the text direction. Allowed values are ltr (Left-To-Right) or rtl (Right-To-Left)}

\item{draggable}{Defines whether the element can be dragged.}

\item{hidden}{Prevents rendering of given element, while keeping child elements, e.g. script elements, active.}

\item{lang}{Defines the language used in the element.}

\item{spellCheck}{Indicates whether spell checking is allowed for the element.}

\item{style}{Defines CSS styles which will override styles previously set.}

\item{tabIndex}{Overrides the browser's default tab order and follows the one specified instead.}

\item{title}{Text to be displayed in a tooltip when hovering over the element.}

\item{fireEvent}{A callback for firing events to dash.}

\item{dashEvents}{}
}
\description{
See \url{https://developer.mozilla.org/en-US/docs/Web/HTML/Element/select}
}
