% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{input}
\alias{input}
\title{input component 
A basic HTML input control for entering text, numbers, or passwords.  Note that checkbox and radio types are supported through the Checklist and RadioItems component. Dates, times, and file uploads are also supported through separate components.}
\usage{
input(id = NULL, value = NULL, placeholder = NULL, style = NULL,
  className = NULL, type = NULL, disabled = NULL, fireEvent = NULL,
  setProps = NULL, dashEvents = NULL)
}
\arguments{
\item{id}{}

\item{value}{The value of the input}

\item{placeholder}{A hint to the user of what can be entered in the control. Note: Do not use the placeholder attribute instead of a Label element, their purposes are different. The Label attribute describes the role of the form element (i.e. it indicates what kind of information is expected), and the placeholder attribute is a hint about the format that the value should take. There are cases in which the placeholder attribute is never displayed to the user, so the form must be understandable without it.}

\item{style}{The input's inline styles}

\item{className}{The class of the input element}

\item{type}{The type of control to render.}

\item{disabled}{If disabled, then the input can not be edited}

\item{fireEvent}{Dash-assigned callback that gets fired when the input changes.}

\item{setProps}{Dash-assigned callback that gets fired when the value changes.}

\item{dashEvents}{}
}
\description{
input component 
A basic HTML input control for entering text, numbers, or passwords.  Note that checkbox and radio types are supported through the Checklist and RadioItems component. Dates, times, and file uploads are also supported through separate components.
}
