% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_snippet.R
\name{install_snippet}
\alias{install_snippet}
\title{Install Dash RStudio snippet}
\usage{
install_snippet()
}
\value{
boolean Whether or not the snippet was installed
}
\description{
Install the Dash code snippet for RStudio, for quickly creating a new Dash
app.
}
