% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{interval}
\alias{interval}
\title{interval component 
A component that repeatedly fires an event ("interval") with a fixed time delay between each event. Interval is good for triggering a component on a recurring basis. The time delay is set with the property "interval" in milliseconds.}
\usage{
interval(id = NULL, interval = NULL, fireEvent = NULL, setProps = NULL,
  dashEvents = NULL)
}
\arguments{
\item{id}{}

\item{interval}{This component will fire an event every `interval` milliseconds with the event name `setInterval`}

\item{fireEvent}{Dash assigned callback}

\item{setProps}{Dash assigned callback}

\item{dashEvents}{}
}
\description{
interval component 
A component that repeatedly fires an event ("interval") with a fixed time delay between each event. Interval is good for triggering a component on a recurring basis. The time delay is set with the property "interval" in milliseconds.
}
