% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{link}
\alias{link}
\title{link component}
\usage{
link(href = NULL, refresh = NULL, className = NULL, style = NULL,
  id = NULL, children = NULL)
}
\arguments{
\item{href}{}

\item{refresh}{}

\item{className}{}

\item{style}{}

\item{id}{}

\item{children}{}
}
\description{

}
