% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{location}
\alias{location}
\title{location component}
\usage{
location(id = NULL, pathname = NULL, refresh = NULL)
}
\arguments{
\item{id}{(required)}

\item{pathname}{}

\item{refresh}{}
}
\description{

}
