% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{markdown}
\alias{markdown}
\title{markdown component 
A component that renders Markdown text as specified by the CommonMark spec.}
\usage{
markdown(id = NULL, className = NULL, containerProps = NULL,
  children = NULL)
}
\arguments{
\item{id}{}

\item{className}{Class name of the container element}

\item{containerProps}{An object containing custom element props to put on the container element such as id or style}

\item{children}{A markdown string that adhreres to the CommonMark spec}
}
\description{
markdown component 
A component that renders Markdown text as specified by the CommonMark spec.
}
