% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{print.Dash}
\alias{print.Dash}
\title{Run a Dash app when explicitly printed to the console}
\usage{
\method{print}{Dash}(x, ...)
}
\description{
Run a Dash app when explicitly printed to the console
}
\keyword{internal}
