% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.dash_component}
\alias{print.dash_component}
\title{Print a dash component object}
\usage{
\method{print}{dash_component}(x, ...)
}
\arguments{
\item{x}{an object of class dash_component}

\item{...}{not currently used}
}
\description{
Print a dash component object
}
\author{
Carson Sievert
}
