% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{run_app}
\alias{run_app}
\title{Run a Dash app}
\usage{
run_app(
  app,
  host = Sys.getenv("DASH_HOST", Sys.getenv("HOST", "127.0.0.1")),
  port = Sys.getenv("DASH_PORT", Sys.getenv("PORT", 8050)),
  browser = interactive()
)
}
\arguments{
\item{app}{A dash application created with \code{\link[=dash_app]{dash_app()}}}

\item{host}{Hostname to run the app.}

\item{port}{Port number to run the app.}

\item{browser}{Whether or not to launch a browser to the app's URL.}
}
\description{
Run a Dash app
}
