% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{runTutorial}
\alias{runTutorial}
\title{Run a dasher example}
\usage{
runTutorial(name, block = FALSE, showcase = interactive(), ...)
}
\arguments{
\item{name}{the example's name.}

\item{block}{whether or not to block the R console.}

\item{showcase}{whether or not to prompt your web browser.}

\item{...}{other arguments passed onto the app's \code{run_server} method}
}
\description{
Run a dasher example
}
\examples{

excel <- runTutorial("excel.R")
excel$server$stop()

}
