% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks-advanced.R
\name{add_callback}
\alias{add_callback}
\title{Add a callback to a Dash app}
\usage{
add_callback(app, outputs, params, callback)
}
\arguments{
\item{app}{A dash application created with \code{\link[=dash_app]{dash_app()}}.}

\item{outputs}{Unnamed list. The \code{output} argument provides the component \code{id}
and \code{property} which will be updated by the callback; a callback can
target one or more outputs (i.e. multiple outputs).}

\item{params}{Unnamed list; provides \link{input} and \link{state} statements, each
with its own defined \code{id} and \code{property}. For pattern-matching callbacks,
the \code{id} field of a component is written in JSON-like syntax and provides
fields that are arbitrary keys which describe the targets of the callback.
See \link{selectors} for more details.}

\item{callback}{Function; must return \link{output} provided \link{input} or \link{state}
arguments. \code{callback} may be any valid R function, or a character string
containing valid JavaScript, or a call to \link{clientsideFunction},
including \code{namespace} and \code{function_name} arguments for a locally served
JavaScript function.}
}
\description{
Add a callback to a Dash app
}
