% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dash.R
\docType{class}
\name{Dash}
\alias{Dash}
\title{Generate a Dash server}
\format{An \link[R6:R6Class]{R6::R6Class} generator object}
\usage{
app <- Dash$new(name = "dash", server = fiery::Fire$new(), url_base_pathname = "/", csrf_protect = TRUE)
}
\description{
Description coming soon
}
\section{Arguments}{

\tabular{lll}{
\code{name} \tab  \tab The name of the Dash application.\cr
\code{server} \tab  \tab The web server which powers the application. Currently only \link[fiery:Fire]{fiery::Fire} objects are allowed.\cr
\code{url_base_pathname} \tab  \tab The base path for locating the Dash application.\cr
\code{csrf_protect} \tab  \tab Whether to protect the Dash application from CSRF vulnerabilities.
\code{serve_locally} \tab  \tab Whether to serve HTML dependencies using local files or via a URL
}
}

\section{Fields}{

\describe{
\item{\code{server}}{A \link[fiery:Fire]{fiery::Fire} object}
}
}

\section{Methods}{

\describe{
\item{\code{layout_set(...)}}{
Set the layout (i.e., user interface).
}
\item{\code{layout_get()}}{
Retrieves the layout.
}
\item{\code{callback(fun, output)}}{
Callback function to execute when relevant inputs change.
}
\item{\code{dependencies_set(dependencies = NULL, section = NULL, priority = NULL)}}{
Adds additional HTML dependencies to your dash application (beyond the 'internal' dependencies).
The \code{dependencies} argument accepts \link[htmltools:htmlDependency]{htmltools::htmlDependency} or
\link[htmltools:htmlDependencies]{htmltools::htmlDependencies}. The \code{section} argument determines whether
your dependencies are placed inside \code{<head>} or \code{<footer>}. The \code{priority}
argument controls the order of the dependencies within a \code{section} (lower
numbers are granted higher priority). If provided, the \code{section} and \code{priority}
arguments should either be of length 1 or the same length as \code{dependencies}.
}
\item{\code{dependencies_get(all = FALSE)}}{
Retrieve (just user-defined or all) HTML dependencies.
}
\item{\code{config_set(suppress_callback_exceptions, permissions_cache_expiry)}}{
Set the config.
}
\item{\code{config_get()}}{
Get the config.
}
\item{\code{run_server(block = TRUE, showcase = FALSE, ...)}}{
Launch the application. See \link[fiery:Fire]{fiery::Fire} methods for a description of the arguments
}
}
}

\examples{

# hello dash!
app <- dasher::Dash$new()
app$run_server(showcase = TRUE)

# use the layout_set method to define your UI
app$layout_set(
  html_h2("Hello Dash"),
  "Here is a slider",
  core_slider()
)
# retrieve the layout with layout_get()
app$layout_get()


}
\seealso{
\url{https://plot.ly/dash/}
}
\author{
Carson Sievert
}
\keyword{datasets}
