% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dash.R
\docType{class}
\name{Dash}
\alias{Dash}
\title{Generate a Dash server}
\format{An \link[R6:R6Class]{R6::R6Class} generator object}
\usage{
app <- Dash$new(
  name = "dash",
  server = fiery::Fire$new(),
  static_folder = NULL,
  serve_locally = TRUE,
  routes_pathname_prefix = '/',
  requests_pathname_prefix = '/'
)
}
\description{
Description coming soon
}
\section{Arguments}{

\tabular{lll}{
\code{name} \tab \tab The name of the Dash application (placed in the \code{<title>}
of the HTML page).\cr
\code{server} \tab \tab The web server used to power the application.
Must be a \link[fiery:Fire]{fiery::Fire} object.\cr
\code{static_folder} \tab \tab A character vector of directories for serving
with the application (the default, \code{NULL}, means don't serve additional
directories). If provided, the names attribute defines corresponding url
path, otherwise it defaults to '/'.\cr
\code{serve_locally} \tab \tab Whether to serve HTML dependencies locally or
remotely (via URL).\cr
\code{routes_pathname_prefix} \tab \tab a prefix applied to the backend routes.\cr
\code{requests_pathname_prefix} \tab \tab a prefix applied to request endpoints
made by Dash's front-end.\cr
\code{suppress_callback_exceptions} \tab \tab Whether to relay warnings about
possible layout mis-specifications when registering a callback. For an
example of when it is ok to suppress these callbacks, see
\code{runTutorial("urls-part-2.R")}
}
}

\section{Fields}{

\describe{
\item{\code{server}}{A cloned (and modified) version of the \link[fiery:Fire]{fiery::Fire} object
provided to the \code{server} argument (various routes will be added which enable
the dasher functionality).}
}
}

\section{Methods}{

\describe{
\item{\code{layout_set(...)}}{
Set the layout (i.e., user interface). The layout should be either a
collection of dash components (e.g., \link{coreSlider}, \link{htmlDiv}, etc) or
a function which returns a collection of components.
}
\item{\code{layout_get(render = TRUE)}}{
Retrieves the layout. If render is \code{TRUE}, and the layout is a function,
the result of the function (rather than the function itself) is returned.
}
\item{\code{callback(func = NULL, output = NULL, .dots = NULL)}}{
A callback function defintion. The \code{func} argument accepts any R function
and \code{output} defines which layout component property should adopt the results
(via an \link{output} object). To determine what events trigger this callback,
provide \link{input} (and/or \link{state}) object(s) (which should reference
layout components) as argument value(s) to \code{func}. When dealing with many
arguments, it may be useful to create them programmatically and supply them
as a list via \code{.dots}.
}
\item{\code{callback_print(func = NULL, output = NULL, .dots = NULL, pre = TRUE, ...)}}{
Like \code{callback}, except the results from \code{func} are printed, captured, and
(optionally) wrapped in a \link{htmlPre} component. This makes it easy to place
R console output inside the \code{output} container. If \code{pre = TRUE} (the default),
additional arguments (\code{...}) are passed onto \link{htmlPre}.
}
\item{\code{callback_png(func = NULL, output = NULL, .dots = NULL, width = NULL, height = NULL, cairo = TRUE, ...)}}{
Like \code{callback}, except any graphics \code{func} may produce are captured and
wrapped into a \link{htmlImg} component. This makes it easy to place results
that an R graphics (png) device would give you inside the \code{output} container.
}
\item{\code{dependencies_set(dependencies = NULL, section = NULL, priority = NULL)}}{
Adds additional HTML dependencies to your dash application (beyond the 'internal' dependencies).
The \code{dependencies} argument accepts \link[htmltools:htmlDependency]{htmltools::htmlDependency} or
\link[htmltools:htmlDependencies]{htmltools::htmlDependencies}. The \code{section} argument determines whether
your dependencies are placed inside \code{<head>} or \code{<footer>}. The \code{priority}
argument controls the order of the dependencies within a \code{section} (lower
numbers are granted higher priority). If provided, the \code{section} and \code{priority}
arguments should either be of length 1 or the same length as \code{dependencies}.
}
\item{\code{dependencies_get(all = FALSE)}}{
Retrieve (just user-defined or all) HTML dependencies.
}
\item{\code{run_server(block = TRUE, showcase = FALSE, ...)}}{
Launch the application. See \link[fiery:Fire]{fiery::Fire} methods for a description of the arguments
}
}
}

\examples{

# hello dash!
app <- dasher::Dash$new()
app$run_server(showcase = TRUE)

# use the layout_set method to define your UI
app$layout_set(
  htmlH2("Hello Dash"),
  "Here is a slider",
  coreSlider()
)
# retrieve the layout with layout_get()
app$layout_get()


}
\seealso{
\url{https://plot.ly/dash/}
}
\author{
Carson Sievert
}
\keyword{datasets}
