% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/css.R
\name{dash_css}
\alias{dash_css}
\title{Dash app stylesheets}
\usage{
dash_css(filenames = "docs-base")
}
\arguments{
\item{filenames}{css filenames to use}
}
\description{
Easily include 'official' Dash stylesheet in your dasher applications.
Any of the
}
\details{
Use the 'official' Dash stylesheet in your dasher applications
\url{https://github.com/plotly/dash-app-stylesheets}.
}
\examples{
library(dasher)

app <- Dash$new()

# One way to include CSS to improve the sytle/look of your app is
# to use htmlLink() to link to a stylesheet. Here we use the dash
# styleguide https://codepen.io/chriddyp/pen/bWLwgP
app$layout_set(
  htmlLink(href = "https://codepen.io/chriddyp/pen/bWLwgP.css", rel = "stylesheet"),
  coreInput(id = 'inputID', value = 'initial value', type = "text"),
  htmlDiv(id = 'outputID')
)

app$callback(
  function(x = input("inputID"), y = input("inputID", "type")) {
    sprintf("You've entered: '\%s' into the '\%s' input control", x, y)
  },
  output("outputID")
)

# app$run_server(showcase = TRUE)


# For better performance, consider placing CSS in the <head> of the HTML
# document (via app$dependencies_set(), which automatically places css in <head>)
app$layout_set(
  coreInput(id = 'inputID', value = 'initial value', type = "text"),
  htmlDiv(id = 'outputID')
)
app$dependencies_set(dash_css())
app
}
