% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/css.R
\name{dash_css}
\alias{dash_css}
\title{Dash app stylesheets}
\usage{
dash_css(filenames = "docs-base")
}
\arguments{
\item{filenames}{css filenames to use}
}
\description{
Use the 'official' Dash stylesheet in your dashR applications
\url{https://github.com/plotly/dash-app-stylesheets}. To learn how
to use the default style ("docs-base"), see \url{https://codepen.io/chriddyp/pen/bWLwgP}
}
\examples{
library(dashR)

app <- Dash$new()

app$layout_set(
  coreInput(id = 'inputID', value = 'initial value', type = "text"),
  htmlDiv(id = 'outputID')
)

app$dependencies_set(dash_css())
app
}
