% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dash-package.R
\docType{package}
\name{dash-package}
\alias{dash}
\alias{dash-package}
\title{An implementation of the dash protocol for authoring reactive web applications}
\description{
Coming soon
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/plotly/dashR}
  \item Report bugs at \url{https://github.com/plotly/dashR/issues}
}

}
\author{
\strong{Maintainer}: Ryan Kyle \email{ryan@plot.ly} [translator]

Authors:
\itemize{
  \item Chris Parmer \email{chris@plot.ly}
  \item Carson Sievert
}

Other contributors:
\itemize{
  \item  Plotly [copyright holder]
}

}
