% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\docType{data}
\name{dependencies}
\alias{dependencies}
\alias{output}
\alias{input}
\alias{state}
\alias{dashNoUpdate}
\alias{ALL}
\alias{ALLSMALLER}
\alias{MATCH}
\title{Input/Output/State definitions}
\format{
An object of class \code{name} of length 1.

An object of class \code{name} of length 1.

An object of class \code{name} of length 1.
}
\usage{
output(id, property)

input(id, property)

state(id, property)

dashNoUpdate()

ALL

ALLSMALLER

MATCH
}
\arguments{
\item{id}{a component id}

\item{property}{the component property to use}
}
\description{
Use in conjunction with the \code{callback()} method from the \link{Dash} class
to define the update logic in your application.
}
\details{
The \code{dashNoUpdate()} function permits application developers to prevent a
single output from updating the layout. It has no formal arguments.

\code{ALL}, \code{ALLSMALLER} and \code{MATCH} are symbols corresponding to the
pattern-matching callback selectors with the same names. These allow you
to write callbacks that respond to or update an arbitrary or dynamic
number of components. For more information, see the \code{callback} section
in \link{dash}.
}
\keyword{datasets}
