% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{dependencies}
\alias{dependencies}
\alias{output}
\alias{input}
\alias{state}
\title{Input/Output/State definitions}
\usage{
output(id = NULL, property = "children")

input(id = NULL, property = "value")

state(id = NULL, property = "value")
}
\arguments{
\item{id}{a component id}

\item{property}{the component property to use}
}
\description{
Use in conjunction with the \code{callback()} method from the \link[dasher:Dash]{dasher::Dash} class
to define the update logic in your application.
}
\author{
Carson Sievert
}
