% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{dependencies}
\alias{dependencies}
\alias{deps_output}
\alias{deps_input}
\alias{deps_state}
\title{Trigger callbacks via dependency definitions}
\usage{
deps_output(component_id = NULL, component_property = NULL)

deps_input(component_id = NULL, component_property = "value")

deps_state(component_id = NULL, component_property = "value")
}
\description{
Use in conjunction with the \code{callback}
}
\author{
Carson Sievert
}
