% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{dependencies}
\alias{dependencies}
\alias{output}
\alias{input}
\alias{state}
\title{Input/Output/State definitions}
\usage{
output(id, property)

input(id, property)

state(id, property)
}
\arguments{
\item{id}{a component id}

\item{property}{the component property to use}
}
\description{
Use in conjunction with the \code{callback()} method from the \link[dashR:Dash]{dashR::Dash} class
to define the update logic in your application.
}
