% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-output-value.R
\name{format_output_value}
\alias{format_output_value}
\title{Format callback output}
\usage{
format_output_value(x, ...)
}
\arguments{
\item{x}{the output value from a callback function}

\item{...}{additional arugments. Currently unused.}
}
\description{
This (generic) function is called on the output value of callback
function(s) -- effectively providing a way for developers to provide
a more official way to leverage custom objects in a dash application.
}
