% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-htmlwidget.R
\name{htmlwidget}
\alias{htmlwidget}
\title{Htmlwidget component}
\usage{
htmlwidget(id = NULL, name = NULL, package = name)
}
\arguments{
\item{id}{a string used for the react container (around the htmlwidget container).}

\item{name}{the name of the htmlwidget}

\item{package}{the htmlwidget's package name}
}
\description{
Htmlwidget component
}
\author{
Carson Sievert
}
