% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.dash_component}
\alias{print.dash_component}
\title{Output a dash component object as JSON}
\usage{
\method{print}{dash_component}(x, ...)
}
\arguments{
\item{x}{an object of class dash_component}

\item{...}{not currently used}
}
\description{
Objects of the \code{dash_component} class support a \code{print} method,
which first processes the nested list object, and then returns
its JSON representation.
}
