% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{wrap_png}
\alias{wrap_png}
\title{Wrap results}
\usage{
wrap_png(func = NULL, width = NULL, height = NULL, cairo = TRUE, ...)
}
\arguments{
\item{func}{a callback function, that when called, produces a graph}

\item{width}{width in pixels}

\item{height}{height in pixels}

\item{cairo}{Use the Cairo package (if installed).}

\item{...}{arguments passed along to the graphics device}
}
\description{
Wrap results
}
\author{
Carson Sievert
}
