% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{wrap_png}
\alias{wrap_png}
\title{Wrap results into a png image}
\usage{
wrap_png(func = NULL, width = NULL, height = NULL, cairo = TRUE, ...)
}
\arguments{
\item{func}{a function, that when called, produces the side-effect of drawing}

\item{width}{width in pixels}

\item{height}{height in pixels}

\item{cairo}{Use the Cairo package (if installed).}

\item{...}{arguments passed along to the graphics device}
}
\description{
Wrap results into a png image
}
\examples{
library(dasher)
library(ggplot2)
data(tips, package = "reshape2")

app <- Dash$new()

span_slider <- function(min = 0.1, max = 1.5, span = 0.1) {
  marks <- seq(min, max, by = span)
  coreSlider(
    id = 'sliderID', min = min, max = max, step = span,
    value = 0.8, marks = setNames(as.list(marks), marks)
  )
}

app$layout_set(
  htmlDiv(id = 'plotID'),
  htmlDiv(
    id = "sliderContainer",
    "Adjust slider to adjust loess span:",
    span_slider(),
    style = list(width = "50\%")
  )
)

tip_plot <- function(span = input("sliderID")) {
  gg <- ggplot(tips, aes(total_bill, tip)) +
    geom_point() +
    geom_smooth(color = "black", span = span)
  print(gg)
}

app$callback(
  wrap_png(tip_plot),
  output(id = 'plotID')
)

app$run_server(showcase = TRUE)
}
\author{
Carson Sievert
}
